// scb-next/web/mfe-cashflow-blotter-origin/src/Cashflow_CN/components/CashflowDetails/MultiExceptions/components/Vostro/interface.ts
export type VostroFormDetails = {
  settlementCode: string;
  ssiSource: string;
  ssiStatus: string;
  effectiveDate: string;
  fmid: string;
  counterpartName: string;
  swiftType: string;
  country: string;
  security: string;
  debitCredit: string;
  settlementMethod: string;
  deliveryMethod: string;
  settlementMeans: string;
  settlementType: string;
  systemCode: string;
  ssiId: string;
  entity: string;
  tradingCurrency: string;
  typology: string;
  ssiType: string;
  settlementAccount: string;
  productGroup: string;
  productFamily: string;
  productType: string;
  coveredPayment: string;
  cmsAccount: string;
  charges: string;
  isThirdpartyPayment: string;

  beneficiaryBic: string;
  beneficiaryName: string;
  beneficiaryName2: string;
  beneficiaryAddress: string;
  beneficiaryAddress1: string;
  beneficiaryAddress2: string;
  beneficiaryCity: string;
  beneficiaryPostcode: string;
  beneficiaryAccount: string;
  beneficiaryCountryIsoCode: string;
  beneficiaryPostCode: string;
  beneficiaryCityName: string;

  accountWithInstitutionBic: string;
  accountWithInstitutionName: string;
  accountWithInstitutionAddress: string;
  accountWithInstitutionCity: string;
  accountWithInstitutionPostcode: string;
  accountWithInstitutionAccount: string;
  accountWithInstitutionCountryIsoCode: string;
  accountWithInstitutionCityName: string;
  accountWithInstitutionAddress1: string;
  accountWithInstitutionAddress2: string;

  receiversCorrespondentBic: string;
  receiversCorrespondentName: string;
  receiversCorrespondentAddress: string;
  receiversCorrespondentCity: string;
  receiversCorrespondentPostcode: string;
  receiversCorrespondentAccount: string;
  receiversCorrespondentCityName: string;
  receiversCorrespondentCountryIsoCode: string;
  receiversCorrespondentAddress1: string;
  receiversCorrespondentAddress2: string;

  intermediaryBic: string;
  intermediaryName: string;
  intermediaryAddress: string;
  intermediaryCity: string;
  intermediaryPostcode: string;
  intermediaryAccount: string;
  intermediaryCityName: string;
  intermediaryCountryIsoCode: string;
  intermediaryAddress1: string;
  intermediaryAddress2: string;

  orderCustomerBic: string;
  orderCustomerName: string;
  orderCustomerAddress: string;
  orderCustomerCity: string;
  orderCustomerPostcode: string;
  orderCustomerAccount: string;
  orderCustomerCityName: string;
  orderCustomerCountryIsoCode: string;
  orderCustomerAddress1: string;
  orderCustomerAddress2: string;

  remittanceInformation1: string;
  remittanceInformation2: string;
  remittanceInformation3: string;
  remittanceInformation4: string;

  senderToReceiver1: string;
  senderToReceiver2: string;
  senderToReceiver3: string;
  senderToReceiver4: string;
  senderToReceiver5: string;
  senderToReceiver6: string;

  eventRowKey: string;
  bookingEntity: string;
  cfiCode: string;
  popDubai: string;
};